+++ Changelog.md
@@ -1,9 +1,12 @@
 # Develop
+
+# v0.14.1
 * Change: Removed dependency on a8s-demo repository.
 * Change: If the user chooses a relative working directory, it is converted to an absolute path before being stored.
 * Change: Bind command now checks if the backend can be reached first.
 * Change: Updated naming conventions. `Consumer Cluster` is now `App Cluster` and `Management Cluster` is now `Control Plane Cluster`.
-* Change: Klutch: don't randomly generate OIDC client secret if it already exists.
+* Bugfix: Klutch: don't randomly generate OIDC client secret if it already exists.
+* Bugfix: Pin Crossplane package version.
 
 # v0.14.0
 * Feature: `klutch` subcommands to automate the local deployment and usage of Klutch with an `a8s` stack.